refactor: dedupe loadNavSourceDocs into _nav-source-docs.ts (T2, supersedes #1604)
