+++ pages/lib/_sidebar-with-defaults.tsx
@@ -38,16 +38,13 @@ import { Island } from "@takazudo/zfb";
 import SidebarTree from "@/components/sidebar-tree";
 import { settings } from "@/config/settings";
 import { defaultLocale, locales, t, type Locale } from "@/config/i18n";
-import { buildLocaleLinks, docsUrl, isDefaultLocaleOnlyPath, navHref, versionedDocsUrl } from "@/utils/base";
+import { buildLocaleLinks, navHref, versionedDocsUrl } from "@/utils/base";
 import {
   isNavVisible,
-  loadCategoryMeta,
-  type CategoryMeta,
   type NavNode,
 } from "@/utils/docs";
 import { buildSidebarForSection } from "@/utils/sidebar";
-import type { DocsEntry } from "@/types/docs-entry";
-import { loadDocs } from "../_data";
+import { loadNavSourceDocs } from "./_nav-source-docs";
 
 export interface SidebarWithDefaultsProps {
   /** Slug of the active doc page, used to highlight the current entry. */
@@ -92,54 +89,6 @@ function remapVersionedHrefs(
     const newHref = versionedDocsUrl(node.slug, version, nodeLang);
     return { ...node, href: newHref, children };
   });
-}
-
-/**
- * Pick the right `loadDocs(...)` collection name and category-meta dir
- * for the active (locale, version) pair, applying the same locale-first
- * + EN-fallback merge that `pages/[locale]/docs/[...slug].tsx` performs
- * in its own `paths()` so the sidebar tree mirrors what those pages
- * enumerate.
- */
-function loadNavSourceDocs(
-  lang: Locale,
-  currentVersion: string | undefined,
-): { docs: DocsEntry[]; categoryMeta: Map<string, CategoryMeta> } {
-  if (currentVersion) {
-    const collectionName = `docs-v-${currentVersion}`;
-    const versionConfig = settings.versions?.find((v) => v.slug === currentVersion);
-    const docs = loadDocs(collectionName).filter((d) => !d.data.draft);
-    const categoryMeta = loadCategoryMeta(versionConfig?.docsDir ?? settings.docsDir);
-    return { docs, categoryMeta };
-  }
-
-  if (lang === defaultLocale) {
-    const docs = loadDocs("docs").filter((d) => !d.data.draft);
-    const categoryMeta = loadCategoryMeta(settings.docsDir);
-    return { docs, categoryMeta };
-  }
-
-  // Non-default locale: locale-first merge with EN fallback so docs the
-  // active locale has not yet translated still appear in the tree.
-  const localeDocs = loadDocs(`docs-${lang}`).filter((d) => !d.data.draft);
-  const baseDocs = loadDocs("docs").filter((d) => !d.data.draft);
-  const localeSlugSet = new Set(localeDocs.map((d) => d.data.slug ?? d.id));
-  const fallbackDocs = baseDocs
-    .filter((d) => !localeSlugSet.has(d.data.slug ?? d.id))
-    .filter((d) => !isDefaultLocaleOnlyPath(docsUrl(d.data.slug ?? d.id)));
-  const allDocs = [...localeDocs, ...fallbackDocs];
-
-  const localeDir =
-    (settings.locales as Record<string, { dir?: string }>)[lang]?.dir ??
-    settings.docsDir;
-  // Base meta first, locale meta wins on overlapping keys — same merge
-  // order [locale]/docs/[...slug].tsx uses in its paths() pass.
-  const categoryMeta = new Map<string, CategoryMeta>([
-    ...loadCategoryMeta(settings.docsDir),
-    ...loadCategoryMeta(localeDir),
-  ]);
-
-  return { docs: allDocs, categoryMeta };
 }
 
 /**
